cap nhat

+++ app/Http/Controllers/admin/MediaController.php
@@ -7,35 +7,35 @@ class MediaController extends Controller {
 	var $_controller="media";	
   	var $_title="Media";
   	var $_icon="icon-settings font-dark";
-	public function getList(){			
-		$strDirUpload=base_path("upload");
-		$arrData = scandir($strDirUpload);		
-		$controller=$this->_controller;	
-		$task="list";
-		$title=$this->_title;
-		$icon=$this->_icon;		
-		
-		$arrPrivilege=getArrPrivilege();
-        $requestControllerAction=$this->_controller."-list";         
-        if(in_array($requestControllerAction,$arrPrivilege)){
-          return view("admin.".$this->_controller.".list",compact("arrData","controller","task","title","icon"));
-        }
-        else{
-          return view("admin.no-access");
-        }
-	}	
+  	public function getList(){			
+  		$strDirUpload=base_path("upload");
+  		$arrData = scandir($strDirUpload);		
+  		$controller=$this->_controller;	
+  		$task="list";
+  		$title=$this->_title;
+  		$icon=$this->_icon;		
+  		
+  		$arrPrivilege=getArrPrivilege();
+  		$requestControllerAction=$this->_controller."-list";         
+  		if(in_array($requestControllerAction,$arrPrivilege)){
+  			return view("admin.".$this->_controller.".list",compact("arrData","controller","task","title","icon"));
+  		}
+  		else{
+  			return view("admin.no-access");
+  		}
+  	}	
 
 	public function getForm($task){		 
 		$controller=$this->_controller;	
 		$title=$this->_title . " : Add new";
 		$icon=$this->_icon; 			
 		$arrPrivilege=getArrPrivilege();
-        $requestControllerAction=$this->_controller."-form";  
-        if(in_array($requestControllerAction, $arrPrivilege)){
-        	return view("admin.".$this->_controller.".form",compact("controller","task","title","icon"));
-        }else{
-        	return view("admin.no-access");
-        }		
+		$requestControllerAction=$this->_controller."-form";  
+		if(in_array($requestControllerAction, $arrPrivilege)){
+			return view("admin.".$this->_controller.".form",compact("controller","task","title","icon"));
+		}else{
+			return view("admin.no-access");
+		}		
 	}
 
 	public function save(){		
@@ -58,23 +58,30 @@ class MediaController extends Controller {
 	
 	public function trash(){	
 		$checked                =   1;
-	    $type_msg               =   "alert-success";
-	    $msg                    =   "Xóa thành công";   								
-		$arrID=@$_POST["cid"];
-		if(count($arrID) == 0){
-			$checked=0;
-		}
-		if($checked==1){				
-			foreach ($arrID as $key => $value) {
-				if(!empty($value)){
-					$pathFile=base_path("upload/".$value);
-			 		if(file_exists($pathFile)){
-						unlink($pathFile);
-					}	
-				}			 	
-	 		}			 
-		}	
-		return redirect()->route("admin.".$this->_controller.".getList");						
+		$type_msg               =   "alert-success";
+		$msg                    =   "Xóa thành công";   
+		$arrPrivilege=getArrPrivilege();
+		$requestControllerAction=$this->_controller."-trash";
+		if(in_array($requestControllerAction,$arrPrivilege)){
+			$arrID=@$_POST["cid"];
+			if(count($arrID) == 0){
+				$checked=0;
+			}
+			if($checked==1){				
+				foreach ($arrID as $key => $value) {
+					if(!empty($value)){
+						$pathFile=base_path("upload/".$value);
+						if(file_exists($pathFile)){
+							unlink($pathFile);
+						}	
+					}			 	
+				}			 
+			}	
+			return redirect()->route("admin.".$this->_controller.".getList");		
+		}else{
+			return view("admin.no-access");
+		}								
+		
 	}	
 	public function deleteItem(Request $request){
 		$id                     =   $request->id;
